+++ lib/pg_trigger.rb
@@ -5,6 +5,7 @@ require_relative "pg_trigger/railtie"
 
 module PgTrigger
   class InvalidTriggerDefinition < StandardError; end
+  class AmbiguousConditionError < StandardError; end
 
   class << self
     attr_accessor :structure_file_path, :schema, :migrations_path, :raise_on_invalid_definition